fix mcp client
